pyproject.toml
[tool.poetry]
name = "restapi_tutorial"
version = "0.1.0"
description = ""
authors = ["JEONGSUJI <wtw9611@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django-extensions = "^2.2.8"
djangorestframework = "^3.11.0"
notebook = "^6.0.3"
django = "<3"
pygments = "^2.5.2"
model-bakery = "^1.1.0"
model_bakery = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
